+++ sdk/src/client/index.ts
@@ -1,0 +1,1 @@
+export { useDirectus } from './use-directus.js';
